Fix the swift up app delegate
